better auth working but not fully integrated
