+++ drizzle/meta/_journal.json
@@ -36,6 +36,13 @@
       "when": 1749021231634,
       "tag": "0004_opposite_dark_phoenix",
       "breakpoints": true
+    },
+    {
+      "idx": 5,
+      "version": "7",
+      "when": 1749081960612,
+      "tag": "0005_freezing_spyke",
+      "breakpoints": true
     }
   ]
 }
