+++ src/lib/auth-client.ts
@@ -1,0 +1,5 @@
+import { createAuthClient } from "better-auth/react";
+export const authClient = createAuthClient({
+  /** The base URL of the server (optional if you're using the same domain) */
+  // baseURL: "http://localhost:3000"
+});
