+++ src/lib/auth.ts
@@ -1,0 +1,22 @@
+import { betterAuth } from "better-auth";
+import { drizzleAdapter } from "better-auth/adapters/drizzle";
+import { db } from "@/server/db"; // your drizzle instance
+import { account, session, user, verification } from "@/server/db/auth-schema";
+
+export const auth = betterAuth({
+  database: drizzleAdapter(db, {
+    provider: "pg", // or "mysql", "sqlite"
+    schema: {
+      verification: verification,
+      account: account,
+      session: session,
+      user: user,
+    },
+  }),
+  socialProviders: {
+    google: {
+      clientId: process.env.GOOGLE_CLIENT_ID as string,
+      clientSecret: process.env.GOOGLE_CLIENT_SECRET as string,
+    },
+  },
+});
